requirements.txt
pyrofork
tgcrypto-pyrofork
motor
dnspython
requests
python-dotenv
